mcp-api-backup-fastapimcp/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
elasticsearch==8.13.0
fastapi-mcp
